chore(agent-ops): clear the stack-introduced checker debt
Every failure cleared here was verified stack-introduced: all six checkers
pass at the stack's merge-base with main (d1f13cc8d), so none of this is
pre-existing repo debt.

max_lines — seven files get their inline `#[cfg(test)] mod tests` extracted
to a sibling `*_tests.rs` via the repo's `#[path]` idiom, which keeps
`super::*` privacy exactly. Three land clean under 600 that way
(store/agent_wakes.rs, transcript_read.rs, wakes/service.rs); four leave a
smaller allowlist entry than the un-extracted file would.  Ten net-new
allowlist entries remain, each marked `agent-ops stack` and each recorded
as deferred debt rather than excused; four existing entries are bumped.
Stale rationale prose on the ratcheted entries is corrected in the same
pass.

session_mutation_admission — the two unclassified handlers are classified.
`sessions_wakes::schedule_agent_wake` is fenced (it arms a wake that queues
a prompt on the watcher session). `subagents::promote_subagent` is
read-like: one indexed UPDATE of `promoted_at`, no session touched, and it
deliberately takes no permit while its neighbour does — a permit would
refuse promotion whenever a workflow held the parent, for a write the
workflow cannot conflict with. That asymmetry is ratcheted in
api/session_admission_tests.rs.

anyharness_boundaries — the four stale `subagents/mcp/*` and
`sessions/delegation.rs` entries are removed and `sessions/replay.rs` is
tightened 2 -> 1. The one real violation is fixed properly:
`api/http/sessions_wakes.rs` imported `AgentWakeReason` straight from the
store, so the wakes module now re-exports it and the handler names the
domain facade. Seven entries are added, three of them documenting
occurrences that only moved under the agent-ops rename; the rest are marked
NET-NEW.

frontend_boundaries — five of the seven violations are self-references
inside `src/domain` written as `#product/domain/...`; they become relative
imports, matching the 300-odd relative imports already there. The other two
were `pending-agent-updates.ts` reaching up out of the Mobile-safe
`src/domain` cell into `lib/domain/delegated-work`. Rather than punch a
hole in that boundary for one file, the module moves to
`lib/domain/chat/composer/`, where the identity vocabulary it needs lives.
Nothing is allowlisted for this checker.

Also clears a stack-introduced `clippy::manual_contains` in spawn_ops.rs.
