+++ apps/packages/product-client/src/components/workspace/chat/input/PendingAgentUpdatesRow.test.tsx
@@ -7,7 +7,7 @@ import {
   derivePendingPromptQueueRow,
   type PendingPromptQueueEntry,
 } from "#product/domain/chats/pending-prompts/pending-prompt-queue";
-import { groupPendingAgentUpdates } from "#product/domain/chats/pending-prompts/pending-agent-updates";
+import { groupPendingAgentUpdates } from "#product/lib/domain/chat/composer/pending-agent-updates";
 
 function agentEntry(seq: number, sessionId: string, label: string): PendingPromptQueueEntry {
   return {
